==> R/04_describe.qmd <==
---
title: "data_descriptive"
format: html
editor: visual
---

```{r}
library("tidyverse")
```

The expression data used in this project was retrieved from Supplementary tables 2, 3, 5, and 17 of [*(Sharma et. al, 2015)*](https://www.nature.com/articles/nn.4160#Sec12)*.* Tables 2, 5, and 17 collect protein expression data normalized as log LFQ intensity, a metric used to compare the abundance of the **same protein** across multiple samples or conditions. Table 3 presents RNA-sequencing data expressed in RPKM (Reads Per Kilobase of transcript per Million mapped reads). In tables 2, 3, and 5, each of the main cell types (neurons, oligodendrocytes, microglia, and astrocytes) has three technical replicated. In case of cultures cells, oligodendrocytes and neurons are also repeated at different numbers of days in vitro (div).

# Data 2

The `clean_2` data frame contains log₂ LFQ protein intensities measured in four purified mouse brain cell types: **astrocytes, microglia, neurons**, and **oligodendrocytes**. LFQ is a metric used to represent the relative abundance of each protein across the different cell-type samples.

Because the LFQ intensities in this dataset are already provided in **log₂ scale**, they can be directly compared between samples without further transformation. To focus the analysis on cell-type–specific protein expression, we keep only the first six columns of the dataset, which correspond to:

-   *Gene name*

-   *Protein name*

-   *Isolated Astrocytes*

-   *Isolated Microglia*

-   *Isolated Neurons*

-   *Isolated Oligodendrocytes*

To evaluate enrichment of each protein in a given cell type, we compute a **reference abundance** using the median log₂ LFQ intensity across all major brain samples. This provides a biologically meaningful baseline that reflects the “whole brain” proteome against which each isolated cell type can be compared. Using this reference, we calculate the **log₂ fold change** for each protein and each cell type:

*log₂ (FC cell type)​= log₂(LFQ cell type)​−median log₂(LFQ)*

This augmented dataset enables identification of **cell-type enriched proteins** and potential **marker candidates**, in line with the analyses presented in Figure 4b of the original publication.

If we compute the proportion of NAs, we get that more than a third of the data is formed by NAs. However, they are not removed at this point, because we re interested in the differences across cell types. When the difference across cell types (log fold change) is NA, is then removed.

```{r}
clean_2 <- read_tsv(file="../data/clean_data/clean_2.tsv")

missing <- clean_2 |> 
  select(where(is.numeric)) |> 
  summarise(across(everything(), ~ mean(is.na(.x))))

missing
```

# Data 3

The data from table 3 is a RNA sequence expression matrix, containing RPKM values for RNA transcriptome data. The data is obtained from purified mouse brain cell types. The `clean_3` data frame contains 11.062 gene entries, across 21 different cell type samples.

```{r}
clean_3 <- read_tsv(file = "../data/clean_data/clean_3.tsv", show_col_types = FALSE)
dirty_3 <- read_tsv(file = "../data/dirty_data/supp_3.tsv", show_col_types = FALSE)
#colnames(clean_3)
#dim(clean_3)


#Important note on removing NAs:
#NAs compose 22.64% of the total values for clean_3, however drop_na() removes ANY rows containing NAs
#Therefore a substantial data loss is obtained and data analysis should be conducted with caution

rows_in_dirty <- nrow(dirty_3) #19501 entries
rows_in_clean <- nrow(clean_3) #11062 entries


pct_lost <- (rows_in_dirty - rows_in_clean) / rows_in_dirty *100

print(
  paste("The total % of lost rows are", round(pct_lost, 2))
)
```

Analysing 0 values within the dataset:

```{r}
zeros_03 <- clean_3 |> 
  select(where(is.numeric)) |> 
  summarise(across(everything(), function(x) mean(x == 0))
  )

zeros_03

#No zero values within the dataset
```

# Data 5

```{r}
clean_5 <- read_tsv(file = "../data/clean_data/clean_5.tsv", show_col_types = FALSE)
head(clean_5)
```

The `clean_5` data frame contains protein expression data of cultured cell types per individual replicate. As in `clean_2` , the data is expressed in log LFQ intensity, allowing us to both validate our replicates, and to determine the preferential expression of each cultured cell type.

For these goals, we are interested only in columns 1-26, while we can discard information about sequence coverage, molecular weight, and protein IDs.

```{r}
# view(clean_5)

dim(clean_5)
colnames(clean_5)[1:26]
```

The data has 9371 genes and 24 samples: 3 replicated for adult microglia, astrocytes, oligodendrocytes at 3 stages (div 1, 2.5, and 4), and neurons at 3 stages (div 5, 10, and 15).

```{r}
zeros <- clean_5 |> 
  select(where(is.numeric)) |> 
  summarise(across(everything(), function(x) mean(x == 0)))

zeros

zeros_long <- zeros |> 
  pivot_longer(everything(), names_to = "sample", values_to = "zero_fraction")

ggplot(zeros_long, aes(x = sample, y = zero_fraction)) +
  geom_col() +
  coord_flip() +
  labs(
    title = "Fraction of Zero Intensities per Sample",
    y = "Proportion of zeros",
    x = "Sample"
  )

```

```{r}
clean_5 <- clean_5 |>
  select(c(1:26))

stat_5 <- clean_5 |> 
  select(where(is.numeric)) |>    # choosing the numeric columns
  pivot_longer(everything(),
               names_to = "columns",
               values_to = "value") |> 
  group_by(columns) |> 
  summarise(
    count = n(),
    mean = mean(value, na.rm = TRUE),
    median = median(value, na.rm = TRUE),
    sd = sd(value, na.rm = TRUE),
    min = min(value, na.rm = TRUE),
    max = max(value, na.rm = TRUE)
  ) 

stat_5
```

```{r}
per_protein_stat <- clean_5 |> 
  rowwise() |> 
  mutate(
    mean = mean(c_across(where(is.numeric)), na.rm = TRUE),
    sd = sd(c_across(where(is.numeric)), na.rm = TRUE)
  ) |> 
  select(gene_names, protein_names, mean, sd)

per_protein_stat

ggplot(per_protein_stat, aes(x = mean)) +
  geom_histogram(bins = 40) +
  labs(
    title = "Distribution of Protein Mean Intensities",
    x = "Mean LFQ Intensity",
    y = "Count"
  )

```

```{r}
table(sub("_[^_]+$", "", names(clean_5)[sapply(clean_5, is.numeric)]))
```

# Data 17

```{r}
clean_17 <- read_tsv(file = "../data/clean_data/clean_17.tsv")
head(clean_17)
```

This data frame currently has a lot of columns which describe the intensity of RNA signal and proteins in different regions of the mouse brain.

We're specially interested in the columns that describe the intensity of proteins. For this, the columns from 14 to 23 contain the right data.

For the PCA, the only needed data is the brain region and the intensity of signal, so the genes, protein names and everything can be dropped for the analysis.

As mentioned earlier, these columns contain next to 0 NA values, so we don't have to drop them.
