==> R/05_analysis_1.qmd <==
---
title: "Density_plot"
format: html
editor: visual
---

# Density plot

##### Important notes

-   The data is extracted from supplementary table 3

    -   The dataset has been cleaned + augmented prior to this script

    -   All data within RPKM columns is already log2 transformed

    -   Means of cell type specific gene expression is already calculated and stored within RNA columns

-   There are 3 overall classes of column names:

    -   Gene name

    -   Cell type

    -   RNA mean

-   Within the cell type category, the following cells are stored

    -   Oligodendrocytes (div1_1-1_3 + div 2_5 1-3 + div 4 1-3)

    -   Adult microglia 1-3

    -   Atrocytes 1-3

    -   Cortical neurons div05 (1-3) + dig 10 (1-3)

-   RPKM (reads per kilobase of transcript per million mapped reads) is the unit used to measure gene expression of RNA

```{r}
library(tidyverse)
```

```{r}
getwd()

#Going back 1 step to open the file
rna_wide <- read_tsv("../data/augmented_data/table_3_log_means.tsv", show_col_types = FALSE)

#Checking the data again
#colnames(rna_data)
#View(rna_data)
```

Data is read. Steps now:

1.  Create a combined column, containing max RPKM across cell types

2.  Tidy dataset (from wide –\> long)

3.  Group cells into 5 groups (cell types + combined)

4.  Assess number of unique genes (per cell type)

5.  Split data (4 cell types + combined)

6.  Plot using ggplot (4 cell types + combined)

## Step 1: Adding max RPKM values

```{r}
rna_wide <- rna_wide |>
  mutate(
    rna_combined = pmax(
      rna_oligo_div1, rna_oligo_div2_5, rna_oligo_div4,
      rna_microglia, rna_astro,
      rna_neuron_div5, rna_neuron_div10
    )
  )
```

## Step 2: Tidying dataset

```{r}
rna_long <- rna_wide |>
  select(gene_name, starts_with("rna_")) |>
  pivot_longer(
    cols = starts_with("rna_"),
    names_to = "cell_type",
    values_to = "log2_rpkm"
  )

#Checking
#View(rna_long)
#The data is now tidy!
```

## Step 3: Group cells

```{r}
rna_grouped <- rna_long |>
  mutate(
    cell_group = case_when(
      str_detect(cell_type, "oligo")    ~ "Oligodendrocytes",
      str_detect(cell_type, "neuron")   ~ "Neurons",
      str_detect(cell_type, "astro")    ~ "Astrocytes",
      str_detect(cell_type, "microglia")~ "Microglia",
      str_detect(cell_type, "combined") ~ "Combined",
      TRUE ~ NA_character_
    )
  ) |>
  filter(!is.na(cell_group))

#Assessing the data
#View(rna_grouped)
```

## Step 4: Assess unique and repetitive gene counts

```{r}
#Checking if genes appear several times within the dataset
#(how many rows within the rna_grouped dataset have the combination of cell + gene)
count_per_gene <- rna_grouped %>% 
  count(cell_group, gene_name) %>% 
  filter(n > 1)

#Display as range:
count_per_gene_summary <- count_per_gene %>% 
  group_by(cell_group) %>% 
  summarise(
    min_n = min(n),
    max_n = max(n),
    .groups = "drop" #drop group after analysis
  )
  
#Extracting to a printable format
count_per_gene_summary %>% 
  rowwise() %>% 
  mutate(msg = paste0(cell_group, ": Counts per gene ranges from ", min_n, " to ", max_n)) %>% 
  pull(msg) %>% 
  cat(sep = "\n")
```

### Saving unique counts

```{r}
# 4) Unique genes per cell_group (for counts)
rna_unique <- rna_grouped |>
  distinct(cell_group, gene_name, .keep_all = TRUE)

n_all <- rna_unique |>
  count(cell_group, name = "n_all")

n_subset <- rna_unique |>
  filter(log2_rpkm >= 1) |>
  count(cell_group, name = "n_subset")

n_counts <- n_all |>
  left_join(n_subset, by = "cell_group")
```

## Step 5: Split data

Done to enable separate plotting

```{r}
#All cell specific values
rna_all_main <- rna_grouped |>
  filter(cell_group != "Combined")

#Extracted values above the threshold of RPKM >= 1:
rna_sub_main <- rna_all_main |>
  filter(log2_rpkm >= 1)

#Combined cell values
rna_all_combined <- rna_grouped |>
  filter(cell_group == "Combined")

#Extract ones above the threshold
rna_sub_combined <- rna_all_combined |>
  filter(log2_rpkm >= 1)

#Count corresponding values
n_counts_main <- n_counts |>
  filter(cell_group != "Combined")

n_counts_combined <- n_counts |>
  filter(cell_group == "Combined")
```

Output: Total number of genes, compared to number of genes within the subset (above the threshold)

```{r}
n_counts_main
```

```{r}
n_counts_combined
```

## Step 6: Plots

#### Saving x axis values

```{r}
#Defining a shared x axis
x_scale <- scale_x_continuous(
  limits = c(-5, 17),
  breaks = seq(-5, 15, by = 5)
)
```

### Cell type specific plot

```{r}
plot_celltypes <- ggplot() +
  geom_density(
    data = rna_all_main,
    aes(x = log2_rpkm, group = cell_group),
    color = "black"
  ) +
  geom_density(
    data = rna_sub_main,
    aes(x = log2_rpkm, group = cell_group),
    color = "blue"
  ) +
  facet_wrap(~ cell_group, nrow = 2, ncol = 2) +
  x_scale +
  labs(
    title = "Gene expression levels (RNA) per Cell Type",
    x = expression("RPKM (log"[2]*")"),
    y = "Density"
  ) +
  theme_minimal() +
  geom_text(
    data = n_counts_main,
    aes(x = 10, y = 0.19, label = format(n_all, big.mark = ",")),
    hjust = 1, vjust = 1, color = "black", size = 3
  ) +
  geom_text(
    data = n_counts_main,
    aes(x = 10, y = 0.17, label = format(n_subset, big.mark = ",")),
    hjust = 1, vjust = 1, color = "blue", size = 3
  )

ggsave("../results/RNA_density_plot_celltype.png", plot_celltypes, dpi = 300,)
```

### Combined plot

```{r}
plot_combined <- ggplot() +
  geom_density(
    data = rna_all_combined,
    aes(x = log2_rpkm),
    color = "black"
  ) +
  geom_density(
    data = rna_sub_combined,
    aes(x = log2_rpkm),
    color = "blue"
  ) +
  x_scale +
  labs(
    title = "Combined cell types",
    x = expression("RPKM (log"[2]*")"),
    y = "Density"
  ) +
  theme_minimal() +
  geom_text(
    data = n_counts_combined,
    aes(x = 10, y = 0.19, label = format(n_all, big.mark = ",")),
    hjust = 1, vjust = 1, color = "black", size = 3
  ) +
  geom_text(
    data = n_counts_combined,
    aes(x = 10, y = 0.17, label = format(n_subset, big.mark = ",")),
    hjust = 1, vjust = 1, color = "blue", size = 3
  )

ggsave("../results/RNA_density_plot_combined.png", plot_combined, dpi = 300)
```
